requirements.txt
scapy==2.5.0
paho-mqtt==1.6.1
netaddr==1.3.0 